feat(router): Make match options optional in isActive
The behavior now matches RouterLinkActive.
