+++ packages/router/src/url_tree.ts
@@ -81,6 +81,28 @@ const paramCompareMap: Record<ParamMatchOptions, ParamCompareFn> = {
 };
 
 /**
+ * The equivalent `IsActiveMatchOptions` options for `isActive` is called with `true`
+ * (exact = true).
+ */
+export const exactMatchOptions: IsActiveMatchOptions = {
+  paths: 'exact',
+  fragment: 'ignored',
+  matrixParams: 'ignored',
+  queryParams: 'exact',
+};
+
+/**
+ * The equivalent `IsActiveMatchOptions` options for `isActive` is called with `false`
+ * (exact = false).
+ */
+export const subsetMatchOptions: IsActiveMatchOptions = {
+  paths: 'subset',
+  fragment: 'ignored',
+  matrixParams: 'ignored',
+  queryParams: 'subset',
+};
+
+/**
  * Returns a computed signal of whether the given url is activated in the Router.
  *
  * As the router state changes, the signal will update to reflect whether the url is active.
@@ -91,14 +113,14 @@ const paramCompareMap: Record<ParamMatchOptions, ParamCompareFn> = {
 export function isActive(
   url: string | UrlTree,
   router: Router,
-  matchOptions: IsActiveMatchOptions,
+  matchOptions?: IsActiveMatchOptions,
 ): Signal<boolean> {
   const urlTree = url instanceof UrlTree ? url : router.parseUrl(url);
   return computed(() =>
     containsTree(
       router.lastSuccessfulNavigation()?.finalUrl ?? new UrlTree(),
       urlTree,
-      matchOptions,
+      matchOptions ?? subsetMatchOptions,
     ),
   );
 }
